src/requirements.txt
flet==0.28.3
httpx==0.28.1
pydantic==2.11.9
pydantic_settings==2.10.1
PyJWT==2.10.1
PyJWT==2.10.1
